requirements.txt
icalendar==5.0.11
requests==2.31.0
python-dateutil==2.8.2
pytest==7.4.4
